+++ config/deploy.rb
@@ -120,6 +120,7 @@ namespace :deamons do
           execute :systemctl, '--user', :start, :gather_vote_account_details
           execute :systemctl, '--user', :start, :process_ping_thing
           execute :systemctl, '--user', :start, :front_stats_update
+          execute :systemctl, '--user', :start, :leader_stats_update
         end
       end
     end
@@ -137,6 +138,7 @@ namespace :deamons do
           execute :systemctl, '--user', :stop, :gather_vote_account_details
           execute :systemctl, '--user', :stop, :process_ping_thing
           execute :systemctl, '--user', :stop, :front_stats_update
+          execute :systemctl, '--user', :stop, :leader_stats_update
         end
       end
     end
@@ -154,6 +156,7 @@ namespace :deamons do
           execute :systemctl, '--user', :restart, :gather_vote_account_details
           execute :systemctl, '--user', :restart, :process_ping_thing
           execute :systemctl, '--user', :restart, :front_stats_update
+          # execute :systemctl, '--user', :restart, :leader_stats_update
         end
       end
     end
