+++ daemons/front_stats_update_daemon.rb
@@ -2,10 +2,7 @@
 
 require_relative "../config/environment"
 
-@leaders_limit = 6
-@fetched_leaders_limit = 20
-@mainnet = "mainnet"
-@networks = [@mainnet, "testnet"].freeze
+@networks = %w[mainnet testnet].freeze
 sleep_time = 5 # seconds
 update_stats_time = 60 # seconds
 @counter = 0
@@ -18,34 +15,6 @@ def get_latest_cluster_stats
 end
 
 @last_cluster_stats = get_latest_cluster_stats
-
-def all_leaders
-  @networks.map do |network|
-    [network, leaders_for_network(network)]
-  end.to_h
-end
-
-def leaders_for_network(network)
-  client = solana_client(network)
-  current_slot = client.get_slot.result
-  leader_accounts = client.get_slot_leaders(current_slot, @fetched_leaders_limit).result
-  leaders = Validator.where(account: leader_accounts)
-
-  leaders_data(leaders).take(@leaders_limit)
-end
-
-def leaders_data(leaders)
-  leaders.map do |leader|
-    { name: leader.name, avatar_url: leader.avatar_url, account: leader.account }
-  end
-end
-
-def solana_client(network)
-  network == @mainnet ? @mainnet_client : @testnet_client
-end
-
-@mainnet_client = SolanaRpcClient.new.mainnet_client
-@testnet_client = SolanaRpcClient.new.testnet_client
 
 loop do
   begin
@@ -61,9 +30,6 @@ loop do
     print parsed_response
     ActionCable.server.broadcast("front_stats_channel", parsed_response)
 
-    leaders = all_leaders
-    print leaders
-    ActionCable.server.broadcast("leaders_channel", leaders)
     sleep(sleep_time)
   rescue => e
     puts e
