+++ daemons/leader_stats_update.service
@@ -1,0 +1,36 @@
+[Unit]
+Description=update leaders component stats with action cable
+After=syslog.target network.target
+
+[Service]
+Type=simple
+WorkingDirectory=/home/deploy/validators.app/current/
+
+ExecStart=/usr/bin/bundle exec /usr/bin/ruby daemons/leader_stats_update_daemon.rb
+
+UMask=0002
+
+# Greatly reduce Ruby memory fragmentation and heap usage
+# https://www.mikeperham.com/2018/04/25/taming-rails-memory-bloat/
+# Environment=MALLOC_ARENA_MAX=2
+# Environment=LD_PRELOAD=/usr/lib/x86_64-linux-gnu/libjemalloc.so.1
+
+Environment=RAILS_ENV=production
+
+# if we crash, restart
+RestartSec=10
+Restart=always
+
+# output goes to /var/log/syslog
+StandardOutput=syslog
+StandardError=syslog
+
+# Limits
+LimitNOFILE=infinity
+LimitMEMLOCK=infinity
+
+# This will default to "bundler" if we don't specify it
+SyslogIdentifier=leader_stats_update_daemon
+
+[Install]
+WantedBy=default.target
